+++ app/src/demo/table/DemoTable.tsx
@@ -1,15 +1,13 @@
 import React, { useCallback, useEffect, useMemo, useState } from 'react';
 import { Person } from '@epam/uui-docs';
-import { DataRowProps, DataRowOptions, cx, useLazyDataSource, useUuiContext, UuiContexts, DataTableRowProps } from '@epam/uui-core';
-import { Presets, ITablePreset, FlexRow } from '@epam/uui';
-import { DataTable, DataTableRow } from '@epam/promo';
+import { DataRowOptions, cx, useLazyDataSource, useUuiContext, UuiContexts, ITablePreset, useTableState } from "@epam/uui-core";
+import { Presets, FlexRow } from '@epam/uui';
+import { DataTable } from '@epam/promo';
 import css from './DemoTable.scss';
-
 import type { TApi } from '../../data';
 import { getFilters, api } from './data';
-import { getColumns } from './columns';
-import { PersonTableFilter, PersonTableRecord, PersonTableRecordId } from './types';
-import { useTableState } from './hooks';
+import { personColumns } from './columns';
+import { PersonTableFilter } from './types';
 import { FilterPanel } from './FilterPanel';
 import { InfoSidebarPanel } from './InfoSidebarPanel';
 import { SlidingPanel } from './SlidingPanel';
@@ -23,7 +21,6 @@ export const DemoTable: React.FC = () => {
 
     const [initialPresets, setInitialPresets] = useState<ITablePreset[]>([]);
     const filters = useMemo(getFilters, []);
-    const columnsSet = useMemo(() => getColumns(filters), []);
 
     useEffect(() => {
         svc.api.presets.getPresets()
@@ -32,20 +29,19 @@ export const DemoTable: React.FC = () => {
     }, []);
 
     const tableStateApi = useTableState({
-        columns: columnsSet.personColumns,
+        columns: personColumns,
         initialPresets: initialPresets,
         onPresetCreate: svc.api.presets.createPreset,
         onPresetUpdate: svc.api.presets.updatePreset,
         onPresetDelete: svc.api.presets.deletePreset,
     });
 
-    const dataSource = useLazyDataSource<PersonTableRecord, string, PersonTableFilter>({
+    const dataSource = useLazyDataSource<Person, number, PersonTableFilter>({
         api,
-        getId: i => JSON.stringify([i.__typename, i.id]),
-        getChildCount: item => item.__typename === 'PersonGroup' ? item.count : null,
+        getId: i => i.id,
     }, []);
 
-    const { current: rowOptions } = React.useRef<DataRowOptions<PersonTableRecord, string>>({
+    const { current: rowOptions } = React.useRef<DataRowOptions<Person, number>>({
         checkbox: { isVisible: true },
         isSelectable: true,
         onClick(rowProps) {
@@ -54,12 +50,7 @@ export const DemoTable: React.FC = () => {
         },
     });
 
-    const renderRow = (props: DataTableRowProps<PersonTableRecord, PersonTableRecordId>) => {
-        const columns = (props.isLoading || props.value?.__typename === 'Person') ? props.columns : columnsSet.groupColumns;
-        return <DataTableRow key={ props.rowKey } { ...props } size='36' columns={ columns } />;
-    };
-
-    const personsDataView = dataSource.useView(tableStateApi.tableState, tableStateApi.onTableStateChange, {
+    const personsDataView = dataSource.useView(tableStateApi.tableState, tableStateApi.setTableState, {
         rowOptions,
         isFoldedByDefault: () => true,
     });
@@ -75,7 +66,7 @@ export const DemoTable: React.FC = () => {
                 <FilterPanel
                     { ...tableStateApi }
                     filters={ filters }
-                    columns={ columnsSet.personColumns }
+                    columns={ personColumns }
                     closePanel={ () => setIsFilterPanelOpened(false) }
                 />
             </SlidingPanel>
@@ -91,11 +82,11 @@ export const DemoTable: React.FC = () => {
                 <DataTable
                     headerTextCase='upper'
                     getRows={ personsDataView.getVisibleRows }
-                    columns={ columnsSet.personColumns }
-                    renderRow={ renderRow }
+                    columns={ personColumns }
+                    filters={ filters }
                     value={ tableStateApi.tableState }
-                    onValueChange={ tableStateApi.onTableStateChange }
-                    showColumnsConfig
+                    onValueChange={ tableStateApi.setTableState }
+                    showColumnsConfig={ true }
                     allowColumnsResizing
                     allowColumnsReordering
                     { ...personsDataView.getListProps() }
@@ -103,7 +94,7 @@ export const DemoTable: React.FC = () => {
             </div>
 
             <InfoSidebarPanel
-                data={ dataSource.getById(JSON.stringify(["Person", tableStateApi.tableState.selectedId?.[1]])) as Person }
+                data={ dataSource.getById(tableStateApi.tableState.selectedId) }
                 isVisible={ isInfoPanelOpened }
                 onClose={ closeInfoPanel }
             />