+++ app/src/demo/table/FilterPanel/FilterPanel.tsx
@@ -1,19 +1,19 @@
 import React from 'react';
-import { DataColumnProps } from '@epam/uui-core';
+import { DataColumnProps, FilterConfig, ITableState } from "@epam/uui-core";
 import { FlexRow, IconButton, ScrollBars, Text, FlexSpacer } from '@epam/uui';
-import { ITableFilter, ITableState } from '../types';
-import { PresetsBlock } from './PresetsBlock';
-import { FiltersBlock } from './FiltersBlock';
-import { ColumnsBlock } from './ColumnsBlock';
 import { ReactComponent as CloseIcon } from '@epam/assets/icons/common/navigation-close-24.svg';
 
-export interface IFilterPanelProps extends ITableState {
-    columns: DataColumnProps<any>[];
-    filters: ITableFilter[];
+import { PresetsBlock } from "./PresetsBlock";
+import { FiltersBlock } from "./FiltersBlock";
+import { ColumnsBlock } from "./ColumnsBlock";
+
+export interface IFilterPanelProps<TFilter extends Record<string, any>> extends ITableState {
+    columns: DataColumnProps[];
+    filters: FilterConfig<TFilter>[];
     closePanel(): void;
 }
 
-const FilterPanel: React.FC<IFilterPanelProps> = props => {
+const FilterPanel = <TFilter extends Record<string, any>>(props: IFilterPanelProps<TFilter>) => {
     return (
         <>
             <FlexRow borderBottom size='48' padding='18'>
@@ -28,18 +28,18 @@ const FilterPanel: React.FC<IFilterPanelProps> = props => {
                     createNewPreset={ props.createNewPreset }
                     isDefaultPresetActive={ props.isDefaultPresetActive }
                     resetToDefault={ props.resetToDefault }
-                    getActivePresetId={ props.getActivePresetId }
+                    activePresetId={ props.activePresetId }
                     hasPresetChanged={ props.hasPresetChanged }
                     choosePreset={ props.choosePreset }
                 />
                 <FiltersBlock
                     filter={ props.tableState.filter }
-                    onFilterChange={ props.onFilterChange }
+                    onFilterChange={ props.setFilter }
                     filters={ props.filters }
                 />
                 <ColumnsBlock
                     columnsConfig={ props.tableState.columnsConfig }
-                    onColumnsConfigChange={ props.onColumnsConfigChange }
+                    onColumnsConfigChange={ props.setColumnsConfig }
                     columns={ props.columns }
                 />
                 { /*<GroupingBlock/>*/ }
@@ -48,4 +48,4 @@ const FilterPanel: React.FC<IFilterPanelProps> = props => {
     );
 };
 
-export default React.memo(FilterPanel);
+export default React.memo(FilterPanel) as typeof FilterPanel;