+++ app/src/demo/table/data/api.ts
@@ -1,8 +1,9 @@
 import { LazyDataSourceApi, normalizeDataQueryFilter } from "@epam/uui";
+import { Person } from "uui-docs";
 import { svc } from "../../../services";
-import { PersonTableFilter, PersonTableRecord, PersonTableRecordId } from "../types";
+import { PersonTableFilter } from "../types";
 
-const mapFilter = <TFilter extends PersonTableFilter>(filter: TFilter): { [TKey in keyof TFilter]: { in: TFilter[TKey] } } => {
+export const mapFilter = <TFilter extends PersonTableFilter>(filter: TFilter): { [TKey in keyof TFilter]: { in: TFilter[TKey] } } => {
     return Object.keys(filter).reduce((acc, key) => {
         return {
             ...acc,
@@ -13,23 +14,11 @@ const mapFilter = <TFilter extends PersonTableFilter>(filter: TFilter): { [TKey 
     }, {} as { [TKey in keyof TFilter]: { in: TFilter[TKey] } });
 };
 
-export const api: LazyDataSourceApi<PersonTableRecord, string, PersonTableFilter> = (request, ctx) => {
+export const api: LazyDataSourceApi<Person, number, PersonTableFilter> = (request, ctx) => {
     const { ids: clientIds, ...rq } = request;
     const groupBy = rq.filter?.groupBy;
 
-     // Currently, DS shouldn't request items by id in Table scenarios. May need to revisit later.
-    const complexIds = clientIds?.map(id => JSON.parse(id));
-
-    if (groupBy && !ctx?.parent) {
-        return svc.api.demo.personGroups({
-            ...rq,
-            filter: { groupBy },
-            search: null,
-            itemsRequest: { filter: rq.filter, search: rq.search },
-        } as any);
-    } else {
-        const parentFilter = ctx?.parent?.id && groupBy ? { [groupBy + 'Id']: ctx.parent.id } : {};
-        const { groupBy: omitGroupBy, ...mappedFilter } = mapFilter(rq.filter || {});
-        return svc.api.demo.persons({ ...rq, filter: { ...mappedFilter, ...parentFilter } });
-    }
+    const parentFilter = ctx?.parent?.id && groupBy ? { [groupBy + 'Id']: ctx.parent.id } : {};
+    const { groupBy: omitGroupBy, ...mappedFilter } = mapFilter(rq.filter || {});
+    return svc.api.demo.persons({ ...rq, filter: { ...mappedFilter, ...parentFilter } });
 };