+++ epam-promo/components/pickers/docs/common.tsx
@@ -1,7 +1,6 @@
 import * as React from 'react';
-import { ArrayDataSource, LazyDataSource, AsyncDataSource } from '@epam/uui-core';
+import { ArrayDataSource, LazyDataSource, AsyncDataSource, PickerBaseOptions } from '@epam/uui-core';
 import { DocBuilder, PropSamplesCreationContext } from '@epam/uui-docs';
-import { PickerBaseOptions } from '@epam/uui-components';
 import { Text } from '../../typography';
 import { DataPickerRow } from '../DataPickerRow';
 import { demoData } from '@epam/uui-docs';