+++ epam-promo/components/tables/DataTable.tsx
@@ -1,7 +1,6 @@
 import * as React from 'react';
-import { PositionValues, VirtualListRenderRowsParams } from '@epam/uui-components';
-import { ColumnsConfig, DataRowProps, useUuiContext, uuiScrollShadows, useColumnsConfig, IEditable, DataTableRowProps,
-    DataTableState, DataTableColumnsConfigOptions, DataSourceListProps, DataColumnProps, cx } from '@epam/uui-core';
+import { PositionValues, VirtualListRenderRowsParams, useColumnsWithFilters } from '@epam/uui-components';
+import { ColumnsConfig, DataRowProps, useUuiContext, uuiScrollShadows, useColumnsConfig, IEditable, DataTableState, DataTableColumnsConfigOptions, DataSourceListProps, DataColumnProps, cx, DataTableRowProps, FilterConfig } from '@epam/uui-core';
 import { ColumnsConfigurationModal, DataTableHeaderRow, DataTableRow, DataTableMods } from './';
 import { VirtualList } from '../';
 import * as css from './DataTable.scss';
@@ -13,11 +12,13 @@ export interface DataTableProps<TItem, TId> extends IEditable<DataTableState>, D
     renderNoResultsBlock?(): React.ReactNode;
     onScroll?(value: PositionValues): void;
     showColumnsConfig?: boolean;
+    filters?: FilterConfig<any>[];
 }
 
 export function DataTable<TItem, TId>(props: React.PropsWithChildren<DataTableProps<TItem, TId> & DataTableMods>) {
     const { uuiModals } = useUuiContext();
-    const { columns, config, defaultConfig } = useColumnsConfig(props.columns, props.value?.columnsConfig);
+    const columnsWithFilters = useColumnsWithFilters(props.columns, props.filters);
+    const { columns, config, defaultConfig } = useColumnsConfig(columnsWithFilters, props.value?.columnsConfig);
 
     const renderRow = React.useCallback((rowProps: DataRowProps<TItem, TId>) => (
         <DataTableRow