+++ loveship/components/datePickers/RangeDatePicker.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import { DropdownBodyProps, RangeDatePickerValue, BaseRangeDatePickerProps, BaseRangeDatePicker } from '@epam/uui-components';
+import { DropdownBodyProps, RangeDatePickerValue, BaseRangeDatePicker } from '@epam/uui-components';
 import { DropdownContainer, FlexRow, TextInput, SizeMod, EditMode, RangeDatePickerBody } from '../index';
-import { IDropdownToggler, uuiMod, IHasRawProps } from '@epam/uui-core';
+import { IDropdownToggler, uuiMod, IHasRawProps, BaseRangeDatePickerProps } from "@epam/uui-core";
 import cx from 'classnames';
 import { TextSettings } from '../../helpers/textLayout';
 import { systemIcons } from '../icons/icons';
@@ -14,7 +14,7 @@ export interface RangeDatePickerProps extends BaseRangeDatePickerProps, SizeMod,
         from?: IHasRawProps<HTMLDivElement>['rawProps'];
         to?: IHasRawProps<HTMLDivElement>['rawProps'];
         body?: IHasRawProps<HTMLDivElement>['rawProps'];
-    }
+    };
 }
 
 export type InputType = 'from' | 'to';