+++ loveship/components/inputs/docs/numericInput.doc.ts
@@ -2,7 +2,7 @@ import { DocBuilder, isReadonlyDoc } from '@epam/uui-docs';
 import { NumericInputProps } from '@epam/uui-components';
 import { NumericInput, NumericInputMods } from '../NumericInput';
 import { iEditable, sizeDoc, textSettingsDoc, isDisabledDoc, iHasPlaceholder, modeDoc, iFormatter } from '../../../docs';
-import { FormContext, GridContext, ResizableContext, DefaultContext } from '../../../docs';
+import { FormContext, GridContext, ResizableContext, DefaultContext, TableContext } from '../../../docs';
 
 const NumericInputDoc = new DocBuilder<NumericInputProps & NumericInputMods>({ name: 'NumericInput', component: NumericInput })
     .implements([iEditable, iHasPlaceholder, sizeDoc, textSettingsDoc, isDisabledDoc, isReadonlyDoc, modeDoc, iFormatter])
@@ -10,6 +10,9 @@ const NumericInputDoc = new DocBuilder<NumericInputProps & NumericInputMods>({ n
     .prop('step', { examples: [2, 5, 10] })
     .prop('min', { examples: [0, 10], isRequired: true })
     .prop('max', { examples: [20, 50], isRequired: true })
-    .withContexts(DefaultContext, FormContext, ResizableContext, GridContext);
+    .prop('mode', { examples: ["form", "cell"] })
+    .prop('align', { examples: ["left", "right"] })
+    .prop('disableArrows', { examples: [true, false] })
+    .withContexts(DefaultContext, FormContext, TableContext, ResizableContext, GridContext);
 
 export = NumericInputDoc;