+++ loveship/skinContext.tsx
@@ -1,7 +1,6 @@
 import * as React from 'react';
 import { ISkin, skinComponent, SkinContextComponentProps, ButtonCoreProps } from '@epam/uui-core';
-import { Checkbox, FlexRow, FlexCell, TextInput, Spinner, Button, ModalWindow, ModalBlocker, ModalFooter, ModalHeader,
-    LabeledInput, IconButton, EpamColor, IconButtonProps } from './components';
+import { Checkbox, FlexRow, FlexCell, TextInput, Spinner, Button, ModalWindow, ModalBlocker, ModalFooter, ModalHeader, LabeledInput, IconButton, EpamColor, IconButtonProps, ColumnPickerFilter, DatePicker, RangeDatePicker } from "./components";
 
 const buttonTypeToColor: { [key: string]: EpamColor } = {
     'success': 'grass',
@@ -35,5 +34,9 @@ export const skinContext: ISkin = {
     ModalFooter: skinComponent(ModalFooter),
     ModalHeader: skinComponent(ModalHeader, props => ({ borderBottom: "night50" as const, ...props })),
     LabeledInput: skinComponent(LabeledInput, props => ({ size: '36' as const, ...props })),
+    // @ts-ignore
+    ColumnPickerFilter: skinComponent(ColumnPickerFilter),
+    DatePicker: skinComponent(DatePicker),
+    RangeDatePicker: skinComponent(RangeDatePicker),
 };