+++ uui-core/data/processing/views/LazyListView.ts
@@ -129,7 +129,7 @@ export class LazyListView<TItem, TId extends DataSourceItemId, TFilter = any> ex
         this.isUpdatePending = false;
 
         let completeReset = false;
-
+        
         if (prevValue == null
             || prevProps == null
             || this.tree == null
@@ -137,6 +137,8 @@ export class LazyListView<TItem, TId extends DataSourceItemId, TFilter = any> ex
             || !isEqual(this.value.sorting, prevValue.sorting)
             || !isEqual(this.value.filter, prevValue.filter)
             || !isEqual(this.props.filter, prevProps.filter)
+            || this.value.page !== prevValue.page
+            || this.value.pageSize !== prevValue.pageSize
         ) {
             this.tree = this.tree || LazyTree.blank<TItem, TId, TFilter>(this.props);
             this.tree = this.tree.clearStructureAndUpdateParams({ ...this.props, api: this.api });