+++ uui-core/types/components/ColumnPickerFilter.ts
@@ -1,0 +1,7 @@
+import { DataSourceItemId } from "../dataSources";
+import { PickerBaseProps } from "./PickerBase";
+
+export type ColumnPickerFilterCoreProps <TItem, TId extends DataSourceItemId> = PickerBaseProps<TItem, TId> & {
+    showSearch?: boolean;
+    close?: () => void;
+};