+++ uui-core/types/tables.ts
@@ -2,17 +2,18 @@ import React, { ReactNode } from 'react';
 import { IEditable, IDisableable, ICanBeInvalid, ICheckable, IDropdownToggler, IHasCX, FlexCellProps } from './props';
 import { SortDirection } from './dataQuery';
 import { DndActorRenderParams, DropParams } from './dnd';
-import { DataRowProps, DataSourceListProps, DataSourceState } from './dataSources';
+import { DataRowProps, DataSourceListProps, DataSourceState, IDataSource } from './dataSources';
 import { ILens } from '..';
 
 export interface DataTableState<TFilter = any> extends DataSourceState<TFilter> {
     columnsConfig?: ColumnsConfig;
+    presetId?: number | null;
 }
 
 export interface DataColumnProps<TItem = any, TId = any, TFilter = any> extends FlexCellProps {
     key: string;
     caption?: React.ReactNode;
-    fix?: 'left' | 'right';
+    fix?: "left" | "right";
     isSortable?: boolean;
     isAlwaysVisible?: boolean;
     isHiddenByDefault?: boolean;
@@ -70,7 +71,7 @@ export type ColumnsConfig = {
 export type IColumnConfig =  {
     isVisible?: boolean;
     order?: string;
-    width?: number | 'auto' | '100%';
+    width?: number | "auto" | "100%";
 };
 
 export type DataTableProps<TItem, TId> = DataSourceListProps & IEditable<DataSourceState> & {
@@ -82,3 +83,50 @@ export type DataTableProps<TItem, TId> = DataSourceListProps & IEditable<DataSou
 export type DataTableConfigModalParams = IEditable<DataSourceState> & {
     columns: DataColumnProps<any, any>[],
 };
+
+type FilterConfigBase<TFilter extends Record<string, any>> = {
+    title: string;
+    field: keyof TFilter;
+    columnKey?: string;
+};
+
+type PickerFilterConfig<TFilter extends Record<string, any>> = FilterConfigBase<TFilter> & {
+    type: "singlePicker" | "multiPicker";
+    dataSource: IDataSource<any, any, any>;
+};
+
+type DatePickerFilterConfig<TFilter extends Record<string, any>> = FilterConfigBase<TFilter> & {
+    type: "datePicker" | "rangeDatePicker";
+};
+
+export type FilterConfig<TFilter extends Record<string, any>> = PickerFilterConfig<TFilter>
+    | DatePickerFilterConfig<TFilter>;
+
+export interface ITablePreset<TFilter = Record<string, any>> {
+    name: string;
+    id: number | null;
+    filter: TFilter;
+    isReadonly?: boolean;
+    columnsConfig: ColumnsConfig;
+}
+
+export interface IPresetsApi {
+    activePresetId: number | null;
+    isDefaultPresetActive: boolean;
+    choosePreset(preset: ITablePreset): void;
+    createNewPreset(name: string): void;
+    resetToDefault(): void;
+    hasPresetChanged(preset: ITablePreset): boolean;
+    duplicatePreset(preset: ITablePreset): void;
+    deletePreset(preset: ITablePreset): void;
+    updatePreset(preset: ITablePreset): void;
+}
+
+export interface ITableState<TFilter = Record<string, any>> extends IPresetsApi {
+    tableState: DataTableState;
+    setTableState(newState: DataTableState): void;
+    setFilter(filter: TFilter): void;
+    setColumnsConfig(columnsConfig: ColumnsConfig): void;
+    presets: ITablePreset<TFilter>[];
+    setPage(page: number): void;
+}